// src/lib/types/contact.d.ts
export type ContactForm = {
	emailSubject: string;
	emailMessage: string;
	website: string;
	subjectOptions: string;

	deadlineType?: string;
	projectName?: string;
	projectDetails?: string;
	budget?: string;
	additionalRequirements?: string;
	deadline?: string;
	otherContact?: string;
	companyName?: string;
	workPosition?: string;
	jobDescription?: string;
	workMode?: string;
	workLocation?: string;
	salaryOffer?: string;
	additionalInformation?: string;
	naturePartnership?: string;
	expectedBenefits?: string;
	organisationType?: string;

	organisationName?: string;
	collaborationDetails?: string;
	expectedOutcomes?: string;
	investmentDetails?: string;
	expectedReturns?: string;

	fullName: string;
	emailAddress?: string;
	mobileNumber?: string;
};